# Dispatcher/caching rewrite to address performance regression