--- a/numba_dpex/core/utils/__init__.py
+++ b/numba_dpex/core/utils/__init__.py
@@ -2,9 +2,13 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+from .caching_utils import build_key, create_func_hash, strip_usm_metadata
 from .suai_helper import SyclUSMArrayInterface, get_info_from_suai
 
 __all__ = [
     "get_info_from_suai",
     "SyclUSMArrayInterface",
+    "create_func_hash",
+    "strip_usm_metadata",
+    "build_key",
 ]