This PR partially addresses the performance regressions described in #886. It contains the following 4 key changes.

(1) ~~The put() function implementation in LRUCache contains a call to get(), which was unnecessary. This PR removes the call to get().~~ Replaced call to get() in LRUCache.put() method with explicit logic to update the linked list keeping track of LRU ordering.

(2) Sha256 hash computation was being performed on every call to build cache key, which in turn was being called on for each dynamic call to a kernel. Instead of computing hash on every dynamic call to a kernel, it can be done once. This PR performs the hash computation for every static instance of a kernel rather than every dynamic instance.

(3) The types of kernel arguments are used as a part of the key when caching. The arguments types were being pre-processed to strip out USM metadata. This functionality was being performed twice for each call, once for building key for kernel module cache and again for kernel bundle cache. This PR changes the logic to perform the pre-processing once per-call.

(4) The function that build the cache key takes variable number of arguments and returns a tuple. The rest of the logic from build key has been moved to dispatcher and func. The key intuition of using variable args is to support the different caches, so far, kernel module cache and kernel bundle cache. Both these caches use different number of keys. (Side note: The build_key function is ideally suited to exist as a static method inside AbstractCache class).

**Effects of optimizations:**
Evaluated the effect of these changes using kmeans implementation from [here](https://github.com/soda-inria/sklearn-numba-dpex).
On Intel ATS GPUs the execution time before the changes is 1.7 seconds. After these changes the execution time reduces to 1.4 seconds. See log below. With numba-dpex 0.19.0 the execution time is 1.1 seconds. These changes partially address the regression introduced after 0.19.0.

Run Log with numba-dpex 0.19.0:
-------------------------------------
python benchmark/kmeans.py
Running Kmeans numba_dpex lloyd GPU ... done in 1.1 s


Run Log with numba-dpex main:
-------------------------------------
python benchmark/kmeans.py
Running Kmeans numba_dpex lloyd GPU ... done in 1.7 s

Run Log with this PR:
------------------------
python benchmark/kmeans.py
Running Kmeans numba_dpex lloyd GPU ... done in 1.4 s

- [x] Have you provided a meaningful PR description?
- [x] Have you added a test, reproducer or referred to an issue with a reproducer?
- [x] Have you tested your changes locally for CPU and GPU devices?
- [x] Have you made sure that new changes do not introduce compiler warnings?
- [x] If this PR is a work in progress, are you filing the PR as a draft?