@diptorupd 2023-02-15
Thanks @adarshyoga for triaging it. 

The `func_hash` is really needed if we run into a situation where the number of cached versions of a kernel exceeds `NUMBA_DPEX_CACHE_SIZE`. Note that the cache limit is per kernel. If the limit is exceeded, then we will start to pickle the kernel to disk. 

Can you try making the func_hash optional based on a flag? That can save us some cycles.